requirements.txt
marinedebrisdetector
up42-py==1.0.0
rasterio==1.2.10
concurrent==1.0
